+++ src/Drama/Internal.hs
@@ -42,11 +42,10 @@ import Prelude hiding (MonadFail)
 #endif
 
 
--- | Monad supporting actor operations. Inspired by Elixir and Erlang's
--- processes.
+-- | Monad supporting actor operations.
 --
 -- @since 0.3.0.0
-newtype Process (msg :: Type -> Type) a = Process (ReaderT (ProcessEnv msg) IO a)
+newtype Actor (msg :: Type -> Type) a = Actor (ReaderT (ActorEnv msg) IO a)
   deriving newtype
     ( Functor
     , Applicative
@@ -62,7 +61,7 @@ newtype Process (msg :: Type -> Type) a = Process (ReaderT (ProcessEnv msg) IO a
 
 
 -- | Wrapper around higher-kinded message types, to make them compatible with
--- the lower-level `Process` machinery.
+-- the lower-level `Actor` machinery.
 --
 -- Higher-kinded message types are defined as GADTs with a type parameter. This
 -- allows specifying the response type for messages.
@@ -73,133 +72,133 @@ data Envelope (msg :: Type -> Type) where
   Call :: MVar res -> msg res -> Envelope msg
 
 
--- | Provided some `ProcessEnv`, convert a `Process` action into an `IO`
+-- | Provided some `ActorEnv`, convert a `Actor` action into an `IO`
 -- action.
 --
 -- @since 0.3.0.0
-runProcess :: MonadIO m => ProcessEnv msg -> Process msg a -> m a
-runProcess processEnv (Process reader) = liftIO $ runReaderT reader processEnv
+runActor :: MonadIO m => ActorEnv msg -> Actor msg a -> m a
+runActor actorEnv (Actor reader) = liftIO $ runReaderT reader actorEnv
 
 
--- | Ambient context provided by the `Process` monad.
+-- | Ambient context provided by the `Actor` monad.
 --
--- Values in `ProcessEnv` are scoped to the current process and cannot be safely
+-- Values in `ActorEnv` are scoped to the current actor and cannot be safely
 -- shared. Functions like `spawn`, `receive`, and `here` use these values as
 -- implicit parameters to avoid leaking internals (and for convenience).
 --
 -- @since 0.3.0.0
-data ProcessEnv msg = ProcessEnv
+data ActorEnv msg = ActorEnv
   { address :: Address msg
-    -- ^ Current process' address.
+    -- ^ Current actor's address.
   , mailbox :: Mailbox msg
-    -- ^ Current process' mailbox.
+    -- ^ Current actor's mailbox.
   , scope :: Scope
-    -- ^ Current process' token used for spawning threads.
+    -- ^ Current actor's token used for spawning threads.
   }
 
 
--- | Address for sending messages to a process. Obtained by running `spawn`,
--- `here`, or `receive` (if another process sends you an address).
+-- | Address for sending messages to an actor. Obtained by running `spawn`,
+-- `here`, or `receive` (if another actor sends you an address).
 --
 -- @since 0.3.0.0
 newtype Address msg = Address (Unagi.InChan (Envelope msg))
 
 
--- | Mailbox where a process receives messages. Cannot be shared with other
--- processes; used implicitly by `receive` and `tryReceive`.
+-- | Mailbox where an actor receives messages. Cannot be shared with other
+-- actors; used implicitly by `receive` and `tryReceive`.
 --
 -- @since 0.3.0.0
 newtype Mailbox msg = Mailbox (Unagi.OutChan (Envelope msg))
 
 
--- | Token delimiting the lifetime of child processes (threads) created by a
--- process.
+-- | Token delimiting the lifetime of child actors (threads) created by an
+-- actor.
 --
 -- @since 0.3.0.0
 newtype Scope = Scope Ki.Scope
 
 
--- | Message type used by processes which do not receive messages.
+-- | Message type used by actors which do not receive messages.
 --
 -- @since 0.3.0.0
 data NoMsg res
 
 
--- | Spawn a child process and return its address.
+-- | Spawn a child actor and return its address.
 --
 -- @since 0.3.0.0
 spawn
-  :: Process msg ()
-  -- ^ Process to spawn
-  -> Process _msg (Address msg)
-  -- ^ Spawned process' address
-spawn process = do
+  :: Actor msg ()
+  -- ^ Actor to spawn
+  -> Actor _msg (Address msg)
+  -- ^ Spawned actor's address
+spawn actor = do
   (inChan, outChan) <- liftIO Unagi.newChan
   let address = Address inChan
   let mailbox = Mailbox outChan
-  spawnImpl address mailbox process
+  spawnImpl address mailbox actor
   pure address
 
 
--- | More efficient version of `spawn`, for processes which receive no messages
+-- | More efficient version of `spawn`, for actors which receive no messages
 -- (@msg ~ `NoMsg`@). See docs for `spawn` for more information.
 --
 -- @since 0.3.0.0
-spawn_ :: Process NoMsg () -> Process msg ()
-spawn_ process = do
+spawn_ :: Actor NoMsg () -> Actor msg ()
+spawn_ actor = do
   let address = Address (error noMsgError)
   let mailbox = Mailbox (error noMsgError)
-  spawnImpl address mailbox process
+  spawnImpl address mailbox actor
 
 
 spawnImpl
   :: Address msg
   -> Mailbox msg
-  -> Process msg ()
-  -> Process _msg ()
-spawnImpl address mailbox process = do
-  Scope kiScope <- Process $ asks scope
-  liftIO $ Ki.fork_ kiScope $ runImpl address mailbox process
+  -> Actor msg ()
+  -> Actor _msg ()
+spawnImpl address mailbox actor = do
+  Scope kiScope <- Actor $ asks scope
+  liftIO $ Ki.fork_ kiScope $ runImpl address mailbox actor
 
 
--- | Block until all child processes have terminated.
+-- | Block until all child actors have terminated.
 --
 -- @since 0.3.0.0
-wait :: Process msg ()
+wait :: Actor msg ()
 wait = do
-  Scope kiScope <- Process $ asks scope
+  Scope kiScope <- Actor $ asks scope
   liftIO $ Ki.wait kiScope
 
 
--- | Return the current process' address.
+-- | Return the current actor's address.
 --
 -- @since 0.3.0.0
-here :: Process msg (Address msg)
-here = Process $ asks address
+here :: Actor msg (Address msg)
+here = Actor $ asks address
 
 
--- | Send a message to another process, expecting no response. Returns
--- immediately without blocking.
+-- | Send a message to another actor, expecting no response. Returns immediately
+-- without blocking.
 --
 -- @since 0.3.0.0
 cast
   :: Address msg
-  -- ^ Process' address
+  -- ^ Actor's address
   -> msg ()
   -- ^ Message to send
-  -> Process _msg ()
+  -> Actor _msg ()
 cast (Address inChan) msg = liftIO $ Unagi.writeChan inChan (Cast msg)
 
 
--- | Send a message to another process, and wait for a response.
+-- | Send a message to another actor, and wait for a response.
 --
 -- @since 0.3.0.0
 call
   :: Address msg
-  -- ^ Process' address
+  -- ^ Actor's address
   -> msg res
   -- ^ Message to send
-  -> Process _msg res
+  -> Actor _msg res
   -- ^ Response
 call (Address inChan) msg = liftIO do
   resMVar <- newEmptyMVar
@@ -212,11 +211,11 @@ call (Address inChan) msg = liftIO do
 --
 -- @since 0.3.0.0
 receive
-  :: (forall res. msg res -> Process msg res)
+  :: (forall res. msg res -> Actor msg res)
   -- ^ Callback function that responds to messages
-  -> Process msg ()
+  -> Actor msg ()
 receive callback = do
-  Mailbox outChan <- Process $ asks mailbox
+  Mailbox outChan <- Actor $ asks mailbox
   envelope <- liftIO $ Unagi.readChan outChan
   case envelope of
     Cast msg ->
@@ -230,11 +229,11 @@ receive callback = do
 --
 -- @since 0.3.0.0
 tryReceive
-  :: (forall res. msg res -> Process msg res)
+  :: (forall res. msg res -> Actor msg res)
   -- ^ Callback function that responds to messages
-  -> Process msg ()
+  -> Actor msg ()
 tryReceive callback = do
-  Mailbox outChan <- Process $ asks mailbox
+  Mailbox outChan <- Actor $ asks mailbox
   (element, _) <- liftIO $ Unagi.tryReadChan outChan
   envelope <- liftIO $ Unagi.tryRead element
   case envelope of
@@ -247,10 +246,10 @@ tryReceive callback = do
       liftIO $ putMVar resMVar res
 
 
--- | Run a top-level process. Intended to be used at the entry point of your
+-- | Run a top-level actor. Intended to be used at the entry point of your
 -- program.
 --
--- If your program is designed with processes in mind, you can use `Process` as
+-- If your program is designed with actors in mind, you can use `Actor` as
 -- your program's base monad:
 --
 -- > main :: IO ()
@@ -261,37 +260,37 @@ tryReceive callback = do
 -- @transformers@ or @mtl@.
 --
 -- @since 0.3.0.0
-run :: MonadIO m => Process msg a -> m a
-run process = do
+run :: MonadIO m => Actor msg a -> m a
+run actor = do
   (inChan, outChan) <- liftIO Unagi.newChan
   let address = Address inChan
   let mailbox = Mailbox outChan
-  runImpl address mailbox process
+  runImpl address mailbox actor
 
 
--- | More efficient version of `run`, for processes which receive no messages
+-- | More efficient version of `run`, for actors which receive no messages
 -- (@msg ~ `NoMsg`@). See docs for `run` for more information.
 --
 -- @since 0.3.0.0
-run_ :: MonadIO m => Process NoMsg a -> m a
-run_ process = do
+run_ :: MonadIO m => Actor NoMsg a -> m a
+run_ actor = do
   let address = Address (error noMsgError)
   let mailbox = Mailbox (error noMsgError)
-  runImpl address mailbox process
+  runImpl address mailbox actor
 
 
-runImpl :: MonadIO m => Address msg -> Mailbox msg -> Process msg a -> m a
-runImpl address mailbox process = do
+runImpl :: MonadIO m => Address msg -> Mailbox msg -> Actor msg a -> m a
+runImpl address mailbox actor = do
   liftIO $ Ki.scoped \kiScope -> do
     let scope = Scope kiScope
-    runProcess ProcessEnv{address, mailbox, scope} process
+    runActor ActorEnv{address, mailbox, scope} actor
 
 
 noMsgError :: String
 noMsgError = unlines . fmap unwords $
   [ ["[!] drama internal error"]
   , []
-  , [ "Attempted to use the address or mailbox of a process which cannot send"
+  , [ "Attempted to use the address or mailbox of a actor which cannot send"
     , "or receive messages (msg ~ NoMsg)."
     ]
   , [ "This should be impossible using non-internal modules!" ]